+++ changelog.md
@@ -17,6 +17,7 @@ Codion Change Log
 - FilterTable bug fixed, now sets initial filter view after columns have been initialized, for the horizontalAlignment.
 - FilterTableCellEditor.Builder.clickCountToStart() added.
 - DefaultFilterTableCellEditor bug fixed, JCheckBox based editor did not paint selection background, added some DefaultTableCellEditor behaviour.
+- DefaultFilterTableCellEditor, stop edit action listeners added.
 ### is.codion.framework.domain
 - EntityType bug fixed, define(List) parameter type changed to List<? extends AttributeDefinition.Builder<?, ?>>.
 ### is.codion.framework.db.rmi